pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vsc-sync"
version = "0.1.0"
description = "A CLI tool to synchronize VSCode-like configurations across multiple editors"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["vscode", "configuration", "sync", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
vsc-sync = "vsc_sync.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/vsc-sync-cli"
Repository = "https://github.com/yourusername/vsc-sync-cli"
Issues = "https://github.com/yourusername/vsc-sync-cli/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vsc_sync"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
target-version = "py38"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "S", "BLE", "FBT", "A", "COM", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "G", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "S101"]  # Line too long, assert usage

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "PLR2004"]  # Allow assert and magic values in tests

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=vsc_sync",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]